Temporary quick-fix for a warning in the linux builds
- libconfig uses 'long long' for its 64 bit lookup/set/get functions,
  rather than a well-defined size integer (such as int64_t). GCC
  doesn't like the use of an int64_t* where a long long* is required,
  especially on linux x86_64, where int64_t is typedeffed as long int
- In future, the function signature will be changed, so we can use
  int64 normally.

+++ src/char/char.c
@@ -5919,7 +5919,7 @@ bool char_config_read_player_new(const char *filename, const struct config_t *co
 #else
 	const char *start_point_setting = "start_point_pre";
 #endif
-	int64 i64 = 0;
+	long long int i64 = 0; // TODO: Temporary until the signature of libconfig->setting_lookup_int64 is changed to take a defined-size int (int64)
 
 	nullpo_retr(false, filename);
 	nullpo_retr(false, config);
